debug

+++ test_codes/data_size_test/data_size_test.cu
@@ -481,7 +481,7 @@ int main(int argc, char **argv)
 		long long int iterations = mod / data_stride;////32 * 32 * 4 / 32 * 2 = 256
 	
 		long long int *CPU_data_in;
-		CPU_data_in = (int*)malloc(sizeof(int) * data_size);
+		CPU_data_in = (long long int*)malloc(sizeof(long long int) * data_size);
 		init_cpu_data(CPU_data_in, data_size, data_stride, mod);
 		
 		long long int reduced_iter = iterations;
